ecoprofiler/requirements.txt
streamlit==1.28.0
pandas==2.0.3
geopandas==0.13.2
plotly==5.15.0
folium==0.14.0
streamlit-folium==0.16.0
requests==2.31.0
SPARQLWrapper==2.0.0
praw==7.7.1
pmaw
textblob==0.17.1
python-dotenv==1.0.0
PyYAML==6.0.1
shapely==2.0.1
matplotlib==3.7.2
numpy==1.24.3
reportlab==4.0.4
sentinelhub==3.10.0
geopy==2.3.0  # For geocoding
networkx==3.1
scripts
pillow